requirements.txt
requests==2.31.0
pandas==2.2.0
pyarrow==15.0.0
fastparquet==2024.2.0
duckdb
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv
tqdm
